+++ airframe-jmx/src/main/scala/wvlet/airframe/jmx/MBeanParameter.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.jmx
+package wvlet.airframe.jmx
 
 import wvlet.surface.{ParameterBase, Surface}
